# Improve performance with mutiple patterns