--- a/package.json
+++ b/package.json
@@ -16,7 +16,7 @@
 		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
 	},
 	"scripts": {
-		"bench": "npm update globby glob-stream fast-glob && node bench.js",
+		"bench": "npm update @globby/main-branch glob-stream fast-glob && node bench.js",
 		"test": "xo && ava && tsd"
 	},
 	"files": [
@@ -66,12 +66,12 @@
 		"slash": "^4.0.0"
 	},
 	"devDependencies": {
+		"@globby/main-branch": "sindresorhus/globby#main",
 		"@types/node": "^16.11.11",
 		"ava": "^3.15.0",
 		"benchmark": "2.1.4",
 		"get-stream": "^6.0.1",
 		"glob-stream": "^7.0.0",
-		"globby": "sindresorhus/globby#main",
 		"rimraf": "^3.0.2",
 		"tsd": "^0.19.0",
 		"typescript": "^4.5.2",