Combine tasks with same options.

Test with this pattern `['**/eslint-plugin-*/package.json', '**/@*/eslint-plugin/package.json']`, [Prettier use a similar pattern to search plugins](https://github.com/prettier/prettier/blob/db73cc8e0c9e3e565cf4f5e9eceaa6b7e90b48e4/src/common/load-plugins.js#L106-L110).

```text
main - sync: 1.992s
working - sync: 990.913ms
main - async: 1.059s
working - async: 490.155ms
main - stream: 973.486ms
working - stream: 595.946ms
```